// generated/src/app/api-models/claims-transmittal.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class ClaimsTransmittal {
  seqApTrans: number;
  seqCltrnId: number;
  sortString: string;
  sourceType: string;
  sourceKey: string;
  holdFlag: string;
  checkDate: Date;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  prePriceOnlyFlag: string;
}